Make sure variable is set before using it.
File.exist? doesn't copy well with nil as a parameter.

Thanks for the patch, kytv!

+++ features/support/hooks.rb
@@ -148,7 +148,7 @@ After('@product') do |scenario|
       STDIN.gets
     end
   else
-    if File.exist?(@video_path) && not($config['CAPTURE_ALL'])
+    if @video_path && File.exist?(@video_path) && not($config['CAPTURE_ALL'])
       FileUtils.rm(@video_path)
     end
   end
